requirements.txt
amqp==5.0.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
billiard==3.6.3.0
celery==5.0.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
djangorestframework==3.12.2
dnspython==1.16.0
eventlet==0.29.1
fsspec==0.8.4
greenlet==0.4.17
idna==2.10
kombu==5.0.2
multidict==5.1.0
numpy==1.19.4
pandas==1.1.4
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
yarl==1.6.3
